Add files via upload
The game of Pac Man is a classic video game created by Namco in 1980 (we will use a
modified web clone created by Lucio Panepinto and modified by Chris Marriott). You will
program an artificial intelligence to play as Pac Man to live as long as possible, collect as many
points as possible and to reach the furthest level possible.