Add files via upload
The popular web game 2048 (designed by Gabriele Cirulli) is a game in which a single player
plays against the randomness of the game. Your task is to create an AI that can play the game
well enough to reliably win. If your AI is winning then try to optimize it to get as high a score as
possible.
You are provided with a large code base including Gabriele Cirulli’s original game simulator with
additions made by Chris Marriott to allow for artificial agent control. A sample agent is included
that has a very simple strategy. It selects the first move in the list [UP, LEFT, RIGHT, DOWN]
that is a legal move (results in a new state). The agent uses a clone of the gameboard to
simulate the results of future moves. You must replace its selectMove method with your own
strategy. I recommend using Expectimax or Monte Carlo simulation. In both cases you will
have to also rely on an evaluation function of the current game state.
Your agent will be evaluated by allowing it to play as many rounds of 2048 it can finish in 10
minutes. There is a trade off between choosing moves quickly versus spending time selecting
the best move. All agents capable of winning the game will be submitted to a friendly in-class
competition for highest score.