Add files via upload
In this assignment you will create an agent controller for an agent in a simple virtual world. The
world is a 2-dimensional continuous space (size 800 x 800 pixels). In the world there are three
types of agents:
● Rocks - a rock is inert unless thrown. If thrown by a player it travels at a fixed speed
(twice the maxspeed of a player) in a single direction until hitting a player, a zombie or a
wall, at which point it stops and falls to the ground (or is caught in the case of a player).
● Zombies - a zombie spawns in a random location once every second. A zombie has a
maximum speed that is randomly selected between quite slow and the max speed of a
player. Zombies ignore rocks but chase players. If a zombie catches a player the player
dies (and becomes a zombie).
● Players - a player is controlled by your AI. The player can pick up at most two rocks and
may throw a rock in any direction once per second. The player can move in any
direction up to a maximum speed.
Approximately 60 times per second your agent will have a chance to select an action. This
action will consist of two components: a direction for movement, and a potential throw action.
You can move and throw at the same time but you can only throw a rock if you have a rock and
only once per second. You should not spend too much time selecting your action as there are
other agents in the simulation as well that need to make their move in this timeframe.